컴포넌트로 분리, Props로 데이터 전달, 조건부 렌더링
